requirements.txt
numpy==1.24.1
Pillow==10.1.0
matplotlib==3.8.3
wordcloud==1.9.3
